Reformat for both app1 and app2 use
