+++ src/main/steps/applicant1/confirm-your-joint-application/form.njk
@@ -1,0 +1,9 @@
+{% extends "common/form/form.njk"%}
+{% from "govuk/components/warning-text/macro.njk" import govukWarningText %}
+
+{% block after_fields %}
+  {{ govukWarningText({
+    text: confirmApplicationIsTrueWarning,
+    iconFallbackText: "Warning"
+  }) }}
+{% endblock %}
