+++ src/main/steps/applicant2/confirm-your-joint-application/form.njk
@@ -1,9 +1,1 @@
-{% extends "common/form/form.njk"%}
-{% from "govuk/components/warning-text/macro.njk" import govukWarningText %}
-
-{% block after_fields %}
-  {{ govukWarningText({
-    text: confirmApplicationIsTrueWarning,
-    iconFallbackText: "Warning"
-  }) }}
-{% endblock %}
+{% include "../../applicant1/confirm-your-joint-application/form.njk" %}
