+++ src/main/steps/applicant2/confirm-your-joint-application/template.njk
@@ -1,37 +1,1 @@
-{% from "govuk/components/summary-list/macro.njk" import govukSummaryList %}
-{% from "govuk/components/button/macro.njk" import govukButton %}
-
-{% extends "common/page.njk" %}
-
-{% block backLink %}{% endblock %}
-
-{% block content %}
-  {% block error_summary %}
-    {% include "common/error/summary.njk" %}
-  {% endblock %}
-
-  <h1 class="govuk-heading-xl">{{ title }}</h1>
-  <h3 class="govuk-body-l">{{ subHeader }}</h3>
-
-  {% for key, section in sections %}
-    {% set rows = getAnswerRows(section, true) %}
-    {% if rows | length %}
-      <h2 class="govuk-heading-m">{{ sectionTitles[section] }}</h2>
-
-      {{ govukSummaryList({
-        classes: 'govuk-!-margin-bottom-9',
-        rows: rows
-      }) }}
-    {% endif %}
-  {% endfor %}
-
-  {% block form %}
-    <div class="govuk-grid-row">
-      <div class="govuk-grid-column-two-thirds">
-        {% include "./form.njk" %}
-      </div>
-    </div>
-  {% endblock %}
-
-  {% include "common/contact-us.njk" %}
-{% endblock %}
+{% include "../../applicant1/confirm-your-joint-application/template.njk" %}
